feat(security): Реализована самодостаточная JWT-аутентификация и авторизация в микросервисах
Тело коммита:
- Настроена генерация JWT (Issuer) в user-module с использованием HS512.
- Настроены user-module и notification-module как Resource Servers (JWT Decoder).
- Реализована логика авторизации (Проверка Владения) для DELETE /api/users/{id}.
- Синхронизирован jwt.secret через Config Server.
- Устранены ошибки сериализации DTO (добавлены геттеры).
- Проверена синхронизация токенов через юнит-тест.

+++ user-module/src/main/java/ru/aidar/controller/UserController.java
@@ -2,6 +2,10 @@ package ru.aidar.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
+import ru.aidar.dto.JwtAuthentificationResponseDto;
 import ru.aidar.dto.UserRequestDto;
 import ru.aidar.dto.UserResponseDto;
 import ru.aidar.service.UserService;
@@ -34,12 +38,20 @@ public class UserController {
         return ResponseEntity.ok(userService.getUserById(id));
     }
 
+    @PostMapping("/register")
+    @ResponseStatus(HttpStatus.CREATED)
+    public JwtAuthentificationResponseDto registerUser(@Valid @RequestBody UserRequestDto userRequestDto) {
+        var registeredUser = userService.registerUser(userRequestDto);
+        return registeredUser;
+    }
+
+    /*
     @PostMapping
     public ResponseEntity<UserResponseDto> createUser(@Valid @RequestBody UserRequestDto userRequestDto) {
         UserResponseDto createdUser = userService.createUser(userRequestDto);
         return new ResponseEntity<>(createdUser, HttpStatus.CREATED);
     }
-
+    */
     @PutMapping("/{id}")
     public ResponseEntity<UserResponseDto> updateUser(@PathVariable Long id, @Valid @RequestBody UserRequestDto userRequestDto) {
         UserResponseDto updatedUser = userService.updateUser(id, userRequestDto);
@@ -47,7 +59,12 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteUser(@PathVariable Long id, @AuthenticationPrincipal Jwt jwt) {
+        Long authenticatedUserId = (Long) jwt.getClaim("userId");
+
+        if (!id.equals(authenticatedUserId)) {
+            throw new AccessDeniedException("User is not authorized to delete this resource.");
+        }
         userService.deleteUser(id);
         return ResponseEntity.noContent().build();
     }
